backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pytesseract
pdf2image
pillow
pdfplumber
surya-ocr
torch